[minor] Extract test utils
